Moved empty message into it's own component to keep SearchAndSort.js clean.
