+++ lib/SearchAndSort/SearchAndSort.js
@@ -25,46 +25,7 @@ import craftLayerUrl from '@folio/stripes-components/util/craftLayerUrl';
 import Notes from '../Notes';
 import css from './SearchAndSort.css';
 import makeResourcesAnalyzer from './ResourcesAnalyzer';
-
-function noRecordsMessage(analyzer, searchTerm, filterPaneIsVisible, toggleFilterPane) {
-  const failure = analyzer.failure();
-  let icon = 'search';
-  let label = `No results found for "${searchTerm}". Please check your spelling and filters.`;
-
-  // No search term entered or filters selected
-  if (!analyzer.loaded()) {
-    icon = filterPaneIsVisible ? 'left-arrow' : null;
-    label = 'Choose a filter or enter search query to show results';
-  }
-
-  // Filters selected but no results
-  if (analyzer.loaded() && !searchTerm) {
-    icon = 'search';
-    label = 'No results found. Please check your filters.';
-  }
-
-  // Loading results
-  if (analyzer.pending()) {
-    icon = 'spinner-ellipsis';
-    label = 'Loading...';
-  }
-
-  // Request failure
-  if (failure) {
-    icon = 'validation-error';
-    label = `Error ${failure.httpStatus}`;
-  }
-
-  return (
-    <div className={css.emptyMessage}>
-      <div className={css.emptyMessageLabelWrap}>
-        { icon && <Icon iconRootClass={css.emptyMessageIcon} icon={icon} />}
-        <span className={css.emptyMessageLabel}>{label}</span>
-      </div>
-      { !filterPaneIsVisible && <Button marginBottom0 buttonClass={css.emptyMessageButton} onClick={toggleFilterPane}>Show filters</Button>}
-    </div>
-  );
-}
+import NoResultsMessage from './components/NoResultsMessage';
 
 class SearchAndSort extends React.Component {
   static contextTypes = {
@@ -536,7 +497,7 @@ class SearchAndSort extends React.Component {
           </div>));
 
     const count = analyzer.totalCount();
-    const message = noRecordsMessage(analyzer, searchTerm, filterPaneIsVisible, this.toggleFilterPane);
+    const message = (<NoResultsMessage analyzer={analyzer} searchTerm={searchTerm} filterPaneIsVisible={filterPaneIsVisible} toggleFilterPane={this.toggleFilterPane} />);
     const sortOrder = this.queryParam('sort') || '';
 
     const createRecordLayer = (!this.props.editRecordComponent ? '' :
